chore: restructure sitemap
