+++ app/blogs/sitemap.ts
@@ -1,0 +1,35 @@
+import { MetadataRoute } from 'next'
+import { allBlogs, allGuides } from 'contentlayer/generated'
+import siteMetadata from '@/data/siteMetadata'
+
+export const dynamic = 'force-static'
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  const siteUrl = siteMetadata.siteUrl
+
+  const blogRoutes = allBlogs
+    .filter((post) => !post.draft && !post?.excludeFromSitemap)
+    .map((post) => ({
+      url: `${siteUrl}/${post.path}/`,
+      lastModified: post.lastmod || post.date,
+      changeFrequency: 'weekly' as const,
+      priority: 0.5,
+    }))
+
+  const guideRoutes = allGuides
+    .filter((post) => !post.draft)
+    .map((post) => ({
+      url: `${siteUrl}/${post.path}/`,
+      lastModified: post.lastmod || post.date,
+      changeFrequency: 'weekly' as const,
+      priority: 0.7,
+    }))
+
+  const staticRoutes = ['blog', 'guides'].map((route) => ({
+    url: `${siteUrl}/${route}/`,
+    lastModified: new Date().toISOString().split('T')[0],
+    changeFrequency: 'weekly' as const,
+  }))
+
+  return [...staticRoutes, ...blogRoutes, ...guideRoutes]
+}
