+++ app/docs/sitemap.ts
@@ -1,0 +1,18 @@
+import { MetadataRoute } from 'next'
+import { allDocs } from 'contentlayer/generated'
+import siteMetadata from '@/data/siteMetadata'
+
+export const dynamic = 'force-static'
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  const siteUrl = siteMetadata.siteUrl
+
+  return allDocs
+    .filter((post) => !post.draft)
+    .map((post) => ({
+      url: `${siteUrl}/${post.path}/`,
+      lastModified: post.lastmod || post.date,
+      changeFrequency: 'weekly' as const,
+      priority: 0.5,
+    }))
+}
